+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 1.26.0 (2023-04-04)
 
 - **ENH: Add a function simplifying footprints**
+- **ENH: Pass the `chunks` keyword to `open_rasterio` in `@path_xarr_dst`**
 
 ## 1.25.0 (2023-03-24)